refactor(core): avilla.core.standards
