+++ avilla/core/abstract/message.py
@@ -5,9 +5,10 @@ from typing import TYPE_CHECKING, cast
 
 from graia.amnesia.message import MessageChain
 
-from .metadata import Metadata
 from avilla.core.platform import Land
 from avilla.core.utilles.selector import Selector
+
+from .metadata import Metadata
 
 if TYPE_CHECKING:
     from datetime import datetime
