+++ avilla/core/protocol.py
@@ -10,9 +10,9 @@ from graia.amnesia.message import MessageChain
 from avilla.core._runtime import ctx_avilla, ctx_protocol
 from avilla.core.abstract.account import AbstractAccount
 from avilla.core.abstract.event import AvillaEvent
+from avilla.core.abstract.trait.context import Namespace
 from avilla.core.platform import Abstract, Land, Platform
 from avilla.core.querier import ProtocolAbstractQueryHandler
-from avilla.core.abstract.trait.context import Namespace
 from avilla.core.utilles.event_parser import AbstractEventParser
 from avilla.core.utilles.message_deserializer import MessageDeserializer
 from avilla.core.utilles.message_serializer import MessageSerializer
