+++ avilla/core/standards/request/event.py
@@ -6,9 +6,8 @@ from typing import TYPE_CHECKING
 from graia.broadcast.entities.dispatcher import BaseDispatcher
 from graia.broadcast.interfaces.dispatcher import DispatcherInterface
 
+from avilla.core.abstract.event import AvillaEvent
 from avilla.core.abstract.request import Request
-
-from ..abstract.event import AvillaEvent
 
 if TYPE_CHECKING:
     from datetime import datetime
